MNT: move piecewise functions to separate file closes #667
