+++ rocketpy/mathutils/__init__.py
@@ -1,7 +1,7 @@
 from .function import (
     Function,
-    PiecewiseFunction,
     funcify_method,
     reset_funcified_methods,
 )
 from .vector_matrix import Matrix, Vector
+from .piecewise_function import PiecewiseFunction
